chore: openstatus regions package
