+++ apps/dashboard/package.json
@@ -39,6 +39,7 @@
     "@openstatus/notification-slack": "workspace:*",
     "@openstatus/notification-webhook": "workspace:*",
     "@openstatus/react": "workspace:*",
+    "@openstatus/regions": "workspace:*",
     "@openstatus/tinybird": "workspace:*",
     "@openstatus/tracker": "workspace:*",
     "@openstatus/upstash": "workspace:*",
