+++ apps/web/src/components/monitor-charts/utils.tsx
@@ -1,4 +1,4 @@
-import { regionDict } from "@openstatus/utils";
+import { regionDict } from "@openstatus/regions";
 
 import type { Period, Quantile } from "@/lib/monitor/utils";
 import type { ResponseGraph } from "@/lib/tb";
