+++ packages/db/src/schema/constants.ts
@@ -1,67 +1,10 @@
+import {
+  ALL_REGIONS,
+  FLY_REGIONS,
+  FREE_FLY_REGIONS,
+  regionDict,
+} from "@openstatus/regions";
 import { z } from "zod";
-
-export const flyRegions = [
-  "ams",
-  "arn",
-  "atl",
-  "bog",
-  "bom",
-  "bos",
-  "cdg",
-  "den",
-  "dfw",
-  "ewr",
-  "eze",
-  "fra",
-  "gdl",
-  "gig",
-  "gru",
-  "hkg",
-  "iad",
-  "jnb",
-  "lax",
-  "lhr",
-  "mad",
-  "mia",
-  "nrt",
-  "ord",
-  "otp",
-  "phx",
-  "qro",
-  "scl",
-  "sjc",
-  "sea",
-  "sin",
-  "syd",
-  "waw",
-  "yul",
-  "yyz",
-] as const;
-
-export const koyebRegions = [
-  "koyeb_fra",
-  "koyeb_was",
-  "koyeb_sin",
-  "koyeb_tyo",
-  "koyeb_par",
-  "koyeb_sfo",
-] as const;
-
-export const railwayRegions = [
-  "railway_europe-west4-drams3a",
-  "railway_us-east4-eqdc4a",
-  "railway_asia-southeast1-eqsg3a",
-  "railway_us-west2",
-] as const;
-
-export const freeFlyRegions = [
-  "iad",
-  "ams",
-  "gru",
-  "syd",
-  "sin",
-  "jnb",
-] as const satisfies (typeof flyRegions)[number][];
 
 export const monitorPeriodicity = [
   "30s",
@@ -73,15 +16,16 @@ export const monitorPeriodicity = [
   "other",
 ] as const;
 
-export const monitorRegions = [
-  ...flyRegions,
-  ...koyebRegions,
-  ...railwayRegions,
-] as const;
+export const availableRegions = ALL_REGIONS.filter(
+  (r) => !regionDict[r as keyof typeof regionDict].deprecated,
+);
 
+export const monitorRegions = ALL_REGIONS;
+export const freeFlyRegions = FREE_FLY_REGIONS;
+export const flyRegions = FLY_REGIONS;
 export const monitorPeriodicitySchema = z.enum(monitorPeriodicity);
-export const monitorRegionSchema = z.enum(monitorRegions);
-export const monitorFlyRegionSchema = z.enum(flyRegions);
+export const monitorRegionSchema = z.enum(ALL_REGIONS);
+export const monitorFlyRegionSchema = z.enum(FLY_REGIONS);
 
 export type MonitorFlyRegion = z.infer<typeof monitorFlyRegionSchema>;
 export type Region = z.infer<typeof monitorRegionSchema>;
