+++ packages/notifications/email/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@openstatus/db": "workspace:*",
     "@openstatus/emails": "workspace:*",
+    "@openstatus/regions": "workspace:*",
     "@openstatus/tinybird": "workspace:*",
     "@openstatus/utils": "workspace:*",
     "@react-email/components": "0.0.36",
