+++ packages/regions/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@openstatus/tsconfig/base.json",
+  "include": ["src", "*.ts"]
+}
